@@ -1,21 +1,19 @@
 cmake_minimum_required(VERSION 2.8.3)
 project(ros_can_nodes)
 
-set_directory_properties(PROPERTIES COMPILE_OPTIONS "-std=c++11;") ##-Wall;-Wextra")
+set_directory_properties(PROPERTIES COMPILE_OPTIONS "-std=c++14;") ##-Wall;-Wextra")
 
 ## Find catkin macros and libraries
 ## if COMPONENTS list like find_package(catkin REQUIRED COMPONENTS xyz)
 ## is used, also find other catkin packages
 find_package(catkin REQUIRED COMPONENTS
-  roscpp
   ros_type_introspection
   std_msgs
-  xmlrpcpp
   topic_tools
 )
 
 ## System dependencies are found with CMake's conventions
-find_package(Boost REQUIRED COMPONENTS system)
+find_package(Boost REQUIRED)
 
 
 ## Uncomment this if the package has a setup.py. This macro ensures
@@ -86,8 +84,7 @@ find_package(Boost REQUIRED COMPONENTS system)
 ## DEPENDS: system dependencies of this project that dependent projects also need
 catkin_package(
     INCLUDE_DIRS include
-    #LIBRARIES ros_can_nodes
-    CATKIN_DEPENDS roscpp std_msgs xmlrpcpp ros_type_introspection include/xmlrpcpp topic_tools
+    CATKIN_DEPENDS std_msgs ros_type_introspection topic_tools
     DEPENDS Boost
 )
 
@@ -117,12 +114,12 @@ include_directories(
 ## Declare a C++ executable
 add_executable(ros_can_nodes
     src/RosCanNode.cpp
+    #src/RosCanNodeTest.cpp
     src/TopicBuffers.cpp
     src/CANHelpers.cpp
     src/CANMsgRouter.cpp
     src/MessageBuffer.cpp
     src/RosCanNodeManager.cpp
-    ##src/main.cpp
     src/ros_node_lib/xmlrpc_manager.cpp
     src/ros_node_lib/network.cpp
     src/ros_node_lib/poll_manager.cpp
@@ -134,11 +131,9 @@ add_executable(ros_can_nodes
     src/ros_node_lib/transport_publisher_link.cpp
     src/ros_node_lib/intraprocess_publisher_link.cpp
     src/ros_node_lib/subscription.cpp
-    src/ros_node_lib/subscriber.cpp
     src/ros_node_lib/subscriber_link.cpp
     src/ros_node_lib/transport_subscriber_link.cpp
     src/ros_node_lib/intraprocess_subscriber_link.cpp
-    src/ros_node_lib/single_subscriber_publisher.cpp
     src/ros_node_lib/subscription_queue.cpp
     src/ros_node_lib/callback_queue.cpp
     src/ros_node_lib/rosout_appender.cpp
