@@ -29,11 +29,19 @@
 #define ROSCAN_SUBSCRIPTION_H
 
 #include "common.h"
+#include "rosdefs.h"
 #include "xmlrpc_manager.h"
+#include "callback_queue_interface.h"
 #include <XmlRpc.h>
 #include <ros/header.h>
-#include <ros/statistics.h>
 #include <ros/transport_hints.h>
+#include <vector>
+#include <set>
+#include <map>
+#include <utility>
+#include <mutex>
+#include <boost/shared_ptr.hpp>
+#include <boost/enable_shared_from_this.hpp>
 
 namespace ros {
 
@@ -53,7 +61,8 @@ namespace roscan {
 // Manages a subscription on a single topic.
 class Subscription : public boost::enable_shared_from_this<Subscription> {
     public:
-        Subscription(const RosCanNodePtr& node, const std::string& name, const std::string& md5sum, const std::string& datatype, const ros::TransportHints& transport_hints);
+        Subscription(const RosCanNodePtr& node, const std::string& name, const std::string& md5sum, const std::string& datatype, const ros::TransportHints& transport_hints)
+            : name_{name}, md5sum_{md5sum}, datatype_{datatype}, nonconst_callbacks_{0}, dropped_{false}, shutting_down_{false}, transport_hints_{transport_hints}, node_{node} {}
         virtual ~Subscription();
 
         // Terminate all our PublisherLinks
@@ -72,20 +81,18 @@ class Subscription : public boost::enable_shared_from_this<Subscription> {
         void addLocalConnection(const PublicationPtr& pub);
 
         // Returns whether this Subscription has been dropped or not
-        bool isDropped() { return dropped_; }
+        bool isDropped() const { return dropped_; }
         XmlRpc::XmlRpcValue getStats();
         void getInfo(XmlRpc::XmlRpcValue& info);
 
-        bool addCallback(const ros::SubscriptionCallbackHelperPtr& helper, const std::string& md5sum, CallbackQueueInterface* queue, int32_t queue_size, const ros::VoidConstPtr& tracked_object, bool allow_concurrent_callbacks);
+        bool addCallback(const ros::SubscriptionCallbackHelperPtr& helper, const std::string& md5sum, CallbackQueueInterface *const queue, const int32_t queue_size, const ros::VoidConstPtr& tracked_object, const bool allow_concurrent_callbacks);
         void removeCallback(const ros::SubscriptionCallbackHelperPtr& helper);
-
-        typedef std::map<std::string, std::string> M_string;
 
         // Called to notify that a new message has arrived from a publisher.
         // Schedules the callback for invokation with the callback queue
-        uint32_t handleMessage(const ros::SerializedMessage& m, bool ser, bool nocopy, const boost::shared_ptr<M_string>& connection_header, const PublisherLinkPtr& link);
+        uint32_t handleMessage(const ros::SerializedMessage& m, const bool ser, const bool nocopy, const boost::shared_ptr<std::map<std::string, std::string>>& connection_header, const PublisherLinkPtr& link);
 
-        const std::string datatype();
+        const std::string datatype() const { return datatype_; }
         const std::string md5sum();
 
         // Removes a subscriber from our list
@@ -98,24 +105,24 @@ class Subscription : public boost::enable_shared_from_this<Subscription> {
         // We'll keep a list of these objects, representing in-progress XMLRPC connections to other nodes.
         class PendingConnection : public ASyncXMLRPCConnection {
             public:
-                PendingConnection(XmlRpc::XmlRpcClient* client, ros::TransportUDPPtr udp_transport, const SubscriptionWPtr& parent, const std::string& remote_uri)
-                    : client_(client), udp_transport_(udp_transport), parent_(parent), remote_uri_(remote_uri) {}
+                PendingConnection(XmlRpc::XmlRpcClient *const client, const ros::TransportUDPPtr& udp_transport, const SubscriptionWPtr& parent, const std::string& remote_uri)
+                    : client_{client}, udp_transport_{udp_transport}, parent_{parent}, remote_uri_{remote_uri} {}
 
-                ~PendingConnection() { delete client_; }
+                ~PendingConnection() override { delete client_; }
 
-                XmlRpc::XmlRpcClient* getClient() const { return client_; }
+                XmlRpc::XmlRpcClient *getClient() const { return client_; }
                 ros::TransportUDPPtr getUDPTransport() const { return udp_transport_; }
 
-                virtual void addToDispatch(XmlRpc::XmlRpcDispatch* disp) {
+                void addToDispatch(XmlRpc::XmlRpcDispatch *const disp) override {
                     disp->addSource(client_, XmlRpc::XmlRpcDispatch::WritableEvent | XmlRpc::XmlRpcDispatch::Exception);
                 }
 
-                virtual void removeFromDispatch(XmlRpc::XmlRpcDispatch* disp) {
+                void removeFromDispatch(XmlRpc::XmlRpcDispatch *const disp) override {
                     disp->removeSource(client_);
                 }
 
-                virtual bool check() {
-                    SubscriptionPtr parent = parent_.lock();
+                bool check() override {
+                    const auto parent = parent_.lock();
                     if (!parent) {
                         return true;
                     }
@@ -125,14 +132,13 @@ class Subscription : public boost::enable_shared_from_this<Subscription> {
                         parent->pendingConnectionDone(boost::dynamic_pointer_cast<PendingConnection>(shared_from_this()), result);
                         return true;
                     }
-
                     return false;
                 }
 
-                const std::string& getRemoteURI() { return remote_uri_; }
+                const std::string& getRemoteURI() const { return remote_uri_; }
 
             private:
-                XmlRpc::XmlRpcClient* client_;
+                XmlRpc::XmlRpcClient *client_;
                 ros::TransportUDPPtr udp_transport_;
                 SubscriptionWPtr parent_;
                 std::string remote_uri_;
@@ -146,15 +152,15 @@ class Subscription : public boost::enable_shared_from_this<Subscription> {
         void headerReceived(const PublisherLinkPtr& link, const ros::Header& h);
 
     private:
-        Subscription(const Subscription&);            // not copyable
-        Subscription& operator=(const Subscription&); // nor assignable
+        Subscription(const Subscription&) = delete; // not copyable
+        void operator=(const Subscription&) = delete; // nor assignable
 
         void dropAllConnections();
 
-        void addPublisherLink(const PublisherLinkPtr& link);
+        void addPublisherLink(const PublisherLinkPtr& link) { publisher_links_.push_back(link); }
 
         struct CallbackInfo {
-            CallbackQueueInterface* callback_queue_;
+            CallbackQueueInterface *callback_queue_;
 
             // Only used if callback_queue_ is non-NULL (NodeHandle API)
             ros::SubscriptionCallbackHelperPtr helper_;
@@ -166,27 +172,27 @@ class Subscription : public boost::enable_shared_from_this<Subscription> {
         typedef std::vector<CallbackInfoPtr> V_CallbackInfo;
 
         std::string name_;
-        boost::mutex md5sum_mutex_;
+        std::mutex md5sum_mutex_;
         std::string md5sum_;
         std::string datatype_;
-        boost::mutex callbacks_mutex_;
+        std::mutex callbacks_mutex_;
         V_CallbackInfo callbacks_;
         uint32_t nonconst_callbacks_;
 
         bool dropped_;
         bool shutting_down_;
-        boost::mutex shutdown_mutex_;
+        std::mutex shutdown_mutex_;
 
         typedef std::set<PendingConnectionPtr> S_PendingConnection;
         S_PendingConnection pending_connections_;
-        boost::mutex pending_connections_mutex_;
+        std::mutex pending_connections_mutex_;
 
         V_PublisherLink publisher_links_;
-        boost::mutex publisher_links_mutex_;
+        std::mutex publisher_links_mutex_;
 
         ros::TransportHints transport_hints_;
 
-        ros::StatisticsLogger statistics_;
+        //ros::StatisticsLogger statistics_;
 
         RosCanNodePtr node_;
 
@@ -200,7 +206,7 @@ class Subscription : public boost::enable_shared_from_this<Subscription> {
         typedef std::map<PublisherLinkPtr, LatchInfo> M_PublisherLinkToLatchInfo;
         M_PublisherLinkToLatchInfo latched_messages_;
 
-        typedef std::vector<std::pair<const std::type_info*, ros::MessageDeserializerPtr>> V_TypeAndDeserializer;
+        typedef std::vector<std::pair<const std::type_info *, ros::MessageDeserializerPtr>> V_TypeAndDeserializer;
         V_TypeAndDeserializer cached_deserializers_;
 };
 
