@@ -28,21 +28,15 @@
 #ifndef ROSCAN_TIMER_MANAGER_H
 #define ROSCAN_TIMER_MANAGER_H
 
-#define BOOST_THREAD_HAS_CONDATTR_SET_CLOCK_MONOTONIC
-
-#include <boost/version.hpp>
-#if BOOST_VERSION < 106100
-// use backported version of boost condition variable, see https://svn.boost.org/trac/boost/ticket/6377
-#include "boost_161_condition_variable.h"
-#else
-#include <boost/thread/condition_variable.hpp>
-#endif
-
-#include "common.h"
 #include "callback_queue_interface.h"
-#include <ros/assert.h>
-#include <ros/file_log.h>
 #include <ros/time.h>
+#include <list>
+#include <functional>
+#include <mutex>
+#include <thread>
+#include <condition_variable>
+#include <chrono>
+#include <boost/make_shared.hpp>
 
 namespace roscan {
 
@@ -53,8 +47,8 @@ class TimerManager {
             int32_t handle;
             D period;
 
-            boost::function<void(const E&)> callback;
-            CallbackQueueInterface* callback_queue;
+            std::function<void(const E&)> callback;
+            CallbackQueueInterface *callback_queue;
 
             ros::WallDuration last_cb_duration;
 
@@ -69,7 +63,7 @@ class TimerManager {
             bool has_tracked_object;
 
             // TODO: atomicize
-            boost::mutex waiting_mutex;
+            std::mutex waiting_mutex;
             uint32_t waiting_callbacks;
 
             bool oneshot;
@@ -87,11 +81,11 @@ class TimerManager {
         TimerManager();
         ~TimerManager();
 
-        int32_t add(const D& period, const boost::function<void(const E&)>& callback, CallbackQueueInterface* callback_queue, const ros::VoidConstPtr& tracked_object, bool oneshot);
-        void remove(int32_t handle);
+        int32_t add(const D& period, const std::function<void(const E&)>& callback, CallbackQueueInterface *callback_queue, const ros::VoidConstPtr& tracked_object, const bool oneshot);
+        void remove(const int32_t handle);
 
-        bool hasPending(int32_t handle);
-        void setPeriod(int32_t handle, const D& period, bool reset = true);
+        bool hasPending(const int32_t handle);
+        void setPeriod(const int32_t handle, const D& period, const bool reset = true);
 
         static TimerManager& global() {
             static TimerManager<T, D, E> global;
@@ -101,46 +95,46 @@ class TimerManager {
     private:
         void threadFunc();
 
-        bool waitingCompare(int32_t lhs, int32_t rhs);
-        TimerInfoPtr findTimer(int32_t handle);
+        bool waitingCompare(const int32_t lhs, const int32_t rhs);
+        TimerInfoPtr findTimer(const int32_t handle);
         void schedule(const TimerInfoPtr& info);
         void updateNext(const TimerInfoPtr& info, const T& current_time);
 
         V_TimerInfo timers_;
-        boost::mutex timers_mutex_;
-        boost::condition_variable timers_cond_;
+        std::mutex timers_mutex_;
+        std::condition_variable timers_cond_;
         volatile bool new_timer_;
 
-        boost::mutex waiting_mutex_;
+        std::mutex waiting_mutex_;
         L_int32 waiting_;
 
         uint32_t id_counter_;
-        boost::mutex id_mutex_;
+        std::mutex id_mutex_;
 
         bool thread_started_;
 
-        boost::thread thread_;
+        std::thread thread_;
 
         bool quit_;
 
         class TimerQueueCallback : public CallbackInterface {
             public:
-                TimerQueueCallback(TimerManager<T, D, E>* parent, const TimerInfoPtr& info, T last_expected, T last_real, T current_expected)
-                    : parent_(parent), info_(info), last_expected_(last_expected), last_real_(last_real), current_expected_(current_expected), called_(false) {
-                    boost::mutex::scoped_lock lock(info->waiting_mutex);
+                TimerQueueCallback(TimerManager<T, D, E> *parent, const TimerInfoPtr& info, T last_expected, T last_real, T current_expected)
+                    : parent_{parent}, info_{info}, last_expected_{last_expected}, last_real_{last_real}, current_expected_{current_expected}, called_{false} {
+                    std::lock_guard<std::mutex> lock{info->waiting_mutex};
                     ++info->waiting_callbacks;
                 }
 
                 ~TimerQueueCallback() {
-                    TimerInfoPtr info = info_.lock();
+                    auto info = info_.lock();
                     if (info) {
-                        boost::mutex::scoped_lock lock(info->waiting_mutex);
+                        std::lock_guard<std::mutex> lock{info->waiting_mutex};
                         --info->waiting_callbacks;
                     }
                 }
 
                 CallResult call() {
-                    TimerInfoPtr info = info_.lock();
+                    auto info = info_.lock();
                     if (!info) {
                         return Invalid;
                     }
@@ -168,17 +162,15 @@ class TimerManager {
                         info->callback(event);
                         ros::SteadyTime cb_end = ros::SteadyTime::now();
                         info->last_cb_duration = cb_end - cb_start;
-
                         info->last_real = event.current_real;
 
                         parent_->schedule(info);
                     }
-
                     return Success;
                 }
 
             private:
-                TimerManager<T, D, E>* parent_;
+                TimerManager<T, D, E> *parent_;
                 TimerInfoWPtr info_;
                 T last_expected_;
                 T last_real_;
@@ -189,24 +181,24 @@ class TimerManager {
 };
 
 template <class T, class D, class E>
-TimerManager<T, D, E>::TimerManager() : new_timer_(false), id_counter_(0), thread_started_(false), quit_(false) {}
+TimerManager<T, D, E>::TimerManager() : new_timer_{false}, id_counter_{0}, thread_started_{false}, quit_{false} {}
 
 template <class T, class D, class E>
 TimerManager<T, D, E>::~TimerManager() {
     quit_ = true;
     {
-        boost::mutex::scoped_lock lock(timers_mutex_);
+        std::lock_guard<std::mutex> lock{timers_mutex_};
         timers_cond_.notify_all();
     }
-    if (thread_started_) {
+    if (thread_.joinable()) {
         thread_.join();
     }
 }
 
 template <class T, class D, class E>
-bool TimerManager<T, D, E>::waitingCompare(int32_t lhs, int32_t rhs) {
-    TimerInfoPtr infol = findTimer(lhs);
-    TimerInfoPtr infor = findTimer(rhs);
+bool TimerManager<T, D, E>::waitingCompare(const int32_t lhs, const int32_t rhs) {
+    auto infol = findTimer(lhs);
+    auto infor = findTimer(rhs);
     if (!infol || !infor) {
         return infol < infor;
     }
@@ -214,41 +206,35 @@ bool TimerManager<T, D, E>::waitingCompare(int32_t lhs, int32_t rhs) {
 }
 
 template <class T, class D, class E>
-typename TimerManager<T, D, E>::TimerInfoPtr TimerManager<T, D, E>::findTimer(int32_t handle) {
-    typename V_TimerInfo::iterator it = timers_.begin();
-    typename V_TimerInfo::iterator end = timers_.end();
-    for (; it != end; ++it) {
-        if ((*it)->handle == handle) {
-            return *it;
-        }
-    }
-    return TimerInfoPtr();
+typename TimerManager<T, D, E>::TimerInfoPtr TimerManager<T, D, E>::findTimer(const int32_t handle) {
+    const auto it = std::find_if(timers_.cbegin(), timers_.cend(), [handle](const auto& info){ return handle == info->handle; });
+    return it != timers_.cend() ? *it : TimerInfoPtr();
 }
 
 template <class T, class D, class E>
-bool TimerManager<T, D, E>::hasPending(int32_t handle) {
-    boost::mutex::scoped_lock lock(timers_mutex_);
-    TimerInfoPtr info = findTimer(handle);
+bool TimerManager<T, D, E>::hasPending(const int32_t handle) {
+    std::lock_guard<std::mutex> lock{timers_mutex_};
+    auto info = findTimer(handle);
 
     if (!info) {
         return false;
     }
 
     if (info->has_tracked_object) {
-        ros::VoidConstPtr tracked = info->tracked_object.lock();
+        auto tracked = info->tracked_object.lock();
         if (!tracked) {
             return false;
         }
     }
 
-    boost::mutex::scoped_lock lock2(info->waiting_mutex);
+    std::lock_guard<std::mutex> lock2{info->waiting_mutex};
     return info->next_expected <= T::now() || info->waiting_callbacks != 0;
 }
 
 template <class T, class D, class E>
-int32_t TimerManager<T, D, E>::add(const D& period, const boost::function<void(const E&)>& callback, CallbackQueueInterface* callback_queue,
-                                   const ros::VoidConstPtr& tracked_object, bool oneshot) {
-    TimerInfoPtr info(boost::make_shared<TimerInfo>());
+int32_t TimerManager<T, D, E>::add(const D& period, const std::function<void(const E&)>& callback, CallbackQueueInterface *callback_queue,
+                                   const ros::VoidConstPtr& tracked_object, const bool oneshot) {
+    TimerInfoPtr info{boost::make_shared<TimerInfo>()};
     info->period = period;
     info->callback = callback;
     info->callback_queue = callback_queue;
@@ -265,23 +251,23 @@ int32_t TimerManager<T, D, E>::add(const D& period, const boost::function<void(c
     }
 
     {
-        boost::mutex::scoped_lock lock(id_mutex_);
+        std::lock_guard<std::mutex> lock{id_mutex_};
         info->handle = id_counter_++;
     }
 
     {
-        boost::mutex::scoped_lock lock(timers_mutex_);
+        std::lock_guard<std::mutex> lock{timers_mutex_};
         timers_.push_back(info);
 
         if (!thread_started_) {
-            thread_ = boost::thread(boost::bind(&TimerManager::threadFunc, this));
+            thread_ = std::thread(&TimerManager::threadFunc, this);
             thread_started_ = true;
         }
 
         {
-            boost::mutex::scoped_lock lock(waiting_mutex_);
+            std::lock_guard<std::mutex> lock{waiting_mutex_};
             waiting_.push_back(info->handle);
-            waiting_.sort(boost::bind(&TimerManager::waitingCompare, this, _1, _2));
+            waiting_.sort([this](const auto x, const auto y){ return this->waitingCompare(x, y); });
         }
 
         new_timer_ = true;
@@ -292,33 +278,25 @@ int32_t TimerManager<T, D, E>::add(const D& period, const boost::function<void(c
 }
 
 template <class T, class D, class E>
-void TimerManager<T, D, E>::remove(int32_t handle) {
-    CallbackQueueInterface* callback_queue = 0;
+void TimerManager<T, D, E>::remove(const int32_t handle) {
+    CallbackQueueInterface *callback_queue = nullptr;
     uint64_t remove_id = 0;
 
     {
-        boost::mutex::scoped_lock lock(timers_mutex_);
+        std::lock_guard<std::mutex> lock{timers_mutex_};
 
-        typename V_TimerInfo::iterator it = timers_.begin();
-        typename V_TimerInfo::iterator end = timers_.end();
-        for (; it != end; ++it) {
-            const TimerInfoPtr& info = *it;
-            if (info->handle == handle) {
-                info->removed = true;
-                callback_queue = info->callback_queue;
-                remove_id = (uint64_t)info.get();
-                timers_.erase(it);
-                break;
-            }
+        const auto it = std::find_if(timers_.cbegin(), timers_.cend(), [handle](const auto& info){ return handle == info->handle; });
+        if (it != timers_.cend()) {
+            (*it)->removed = true;
+            callback_queue = (*it)->callback_queue;
+            remove_id = (uint64_t)it->get();
+            timers_.erase(it);
         }
 
         {
-            boost::mutex::scoped_lock lock2(waiting_mutex_);
+            std::lock_guard<std::mutex> lock2{waiting_mutex_};
             // Remove from the waiting list if it's in it
-            L_int32::iterator it = std::find(waiting_.begin(), waiting_.end(), handle);
-            if (it != waiting_.end()) {
-                waiting_.erase(it);
-            }
+            waiting_.remove(handle);
         }
     }
 
@@ -329,7 +307,7 @@ void TimerManager<T, D, E>::remove(int32_t handle) {
 
 template <class T, class D, class E>
 void TimerManager<T, D, E>::schedule(const TimerInfoPtr& info) {
-    boost::mutex::scoped_lock lock(timers_mutex_);
+    std::lock_guard<std::mutex> lock{timers_mutex_};
 
     if (info->removed) {
         return;
@@ -337,11 +315,11 @@ void TimerManager<T, D, E>::schedule(const TimerInfoPtr& info) {
 
     updateNext(info, T::now());
     {
-        boost::mutex::scoped_lock lock(waiting_mutex_);
+        std::lock_guard<std::mutex> lock{waiting_mutex_};
 
         waiting_.push_back(info->handle);
         // waitingCompare requires a lock on the timers_mutex_
-        waiting_.sort(boost::bind(&TimerManager::waitingCompare, this, _1, _2));
+        waiting_.sort([this](const auto x, const auto y){ return this->waitingCompare(x, y); });
     }
 
     new_timer_ = true;
@@ -363,23 +341,22 @@ void TimerManager<T, D, E>::updateNext(const TimerInfoPtr& info, const T& curren
 
         // detect time jumping forward, as well as callbacks that are too slow
         if (info->next_expected + info->period < current_time) {
-            ROS_DEBUG("Time jumped forward by [%f] for timer of period [%f], resetting timer (current=%f, next_expected=%f)", (current_time - info->next_expected).toSec(), info->period.toSec(), current_time.toSec(), info->next_expected.toSec());
             info->next_expected = current_time;
         }
     }
 }
 
 template <class T, class D, class E>
-void TimerManager<T, D, E>::setPeriod(int32_t handle, const D& period, bool reset) {
-    boost::mutex::scoped_lock lock(timers_mutex_);
-    TimerInfoPtr info = findTimer(handle);
+void TimerManager<T, D, E>::setPeriod(const int32_t handle, const D& period, const bool reset) {
+    std::lock_guard<std::mutex> lock{timers_mutex_};
+    auto info = findTimer(handle);
 
     if (!info) {
         return;
     }
 
     {
-        boost::mutex::scoped_lock lock(waiting_mutex_);
+        std::lock_guard<std::mutex> lock{waiting_mutex_};
 
         if (reset) {
             info->next_expected = T::now() + period;
@@ -403,7 +380,7 @@ void TimerManager<T, D, E>::setPeriod(int32_t handle, const D& period, bool rese
         // In this case, let next_expected be updated only in updateNext
 
         info->period = period;
-        waiting_.sort(boost::bind(&TimerManager::waitingCompare, this, _1, _2));
+        waiting_.sort([this](const auto x, const auto y){ return this->waitingCompare(x, y); });
     }
 
     new_timer_ = true;
@@ -416,19 +393,13 @@ void TimerManager<T, D, E>::threadFunc() {
     while (!quit_) {
         T sleep_end;
 
-        boost::mutex::scoped_lock lock(timers_mutex_);
+        std::unique_lock<std::mutex> lock{timers_mutex_};
 
         // detect time jumping backwards
         if (T::now() < current) {
-            ROSCPP_LOG_DEBUG("Time jumped backward, resetting timers");
-
             current = T::now();
 
-            typename V_TimerInfo::iterator it = timers_.begin();
-            typename V_TimerInfo::iterator end = timers_.end();
-            for (; it != end; ++it) {
-                const TimerInfoPtr& info = *it;
-
+            for (const auto& info: timers_) {
                 // Timer may have been added after the time jump, so also check if time has jumped past its last call time
                 if (current < info->last_expected) {
                     info->last_expected = current;
@@ -440,18 +411,17 @@ void TimerManager<T, D, E>::threadFunc() {
         current = T::now();
 
         {
-            boost::mutex::scoped_lock waitlock(waiting_mutex_);
+            std::lock_guard<std::mutex> waitlock{waiting_mutex_};
 
             if (waiting_.empty()) {
                 sleep_end = current + D(0.1);
             } else {
-                TimerInfoPtr info = findTimer(waiting_.front());
+                auto info = findTimer(waiting_.front());
 
                 while (!waiting_.empty() && info && info->next_expected <= current) {
                     current = T::now();
 
-                    //ROS_DEBUG("Scheduling timer callback for timer [%d] of period [%f], [%f] off expected", info->handle, info->period.toSec(), (current - info->next_expected).toSec());
-                    CallbackInterfacePtr cb(boost::make_shared<TimerQueueCallback>(this, info, info->last_expected, info->last_real, info->next_expected));
+                    CallbackInterfacePtr cb{boost::make_shared<TimerQueueCallback>(this, info, info->last_expected, info->last_real, info->next_expected)};
                     info->callback_queue->addCallback(cb, (uint64_t)info.get());
 
                     waiting_.pop_front();
@@ -459,10 +429,8 @@ void TimerManager<T, D, E>::threadFunc() {
                     if (waiting_.empty()) {
                         break;
                     }
-
                     info = findTimer(waiting_.front());
                 }
-
                 if (info) {
                     sleep_end = info->next_expected;
                 }
@@ -473,7 +441,6 @@ void TimerManager<T, D, E>::threadFunc() {
             // detect backwards jumps in time
 
             if (T::now() < current) {
-                ROSCPP_LOG_DEBUG("Time jumped backwards, breaking out of sleep");
                 break;
             }
 
@@ -486,15 +453,14 @@ void TimerManager<T, D, E>::threadFunc() {
             // If we're on simulation time we need to check now() against sleep_end more often than on system time,
             // since simulation time may be running faster than real time.
             if (!T::isSystemTime()) {
-                timers_cond_.timed_wait(lock, boost::posix_time::milliseconds(1));
+                timers_cond_.wait_for(lock, std::chrono::milliseconds(1));
             } else {
                 // On system time we can simply sleep for the rest of the wait time, since anything else requiring processing will
                 // signal the condition variable
-                int32_t remaining_time = std::max((int32_t)((sleep_end - current).toSec() * 1000.0f), 1);
-                timers_cond_.timed_wait(lock, boost::posix_time::milliseconds(remaining_time));
+                int64_t remaining_time = std::max<int64_t>((sleep_end - current).toSec() * 1000.0f, 1);
+                timers_cond_.wait_for(lock, std::chrono::milliseconds(remaining_time));
             }
         }
-
         new_timer_ = false;
     }
 }
