@@ -7,29 +7,29 @@ RosCanNodeManager& RosCanNodeManager::instance() {
     return instance;
 }
 
-roscan::RosCanNodePtr RosCanNodeManager::getNode(uint8_t id) {
-    boost::mutex::scoped_lock nodeListLock(nodeListMutex);
+roscan::RosCanNodePtr RosCanNodeManager::getNode(const uint8_t id) {
+    std::lock_guard<std::mutex> nodeListLock{nodeListMutex};
     return nodeList[id];
 }
 
-uint8_t RosCanNodeManager::registerNode(std::string& name, uint8_t hashName) {
+uint8_t RosCanNodeManager::registerNode(const std::string& name, const uint8_t hashName) {
     uint8_t index = 0;
     //get the first id
 
     {
-        boost::mutex::scoped_lock nodeListLock(nodeListMutex);
+        std::lock_guard<std::mutex> nodeListLock{nodeListMutex};
 
-        while (index < MAX_NODES && nodeList[index] != NULL) {
-            index++;
+        while (index < MAX_NODES && nodeList[index] != nullptr) {
+            ++index;
             if (index >= MAX_NODES) {
-                std::cout << "No available ids" << '\n';
+                std::cout << "No available ids\n";
                 break;
             }
         }
 
         if (index < MAX_NODES) {
             //TODO: take name from frame data
-            nodeList[index].reset(new roscan::RosCanNode(name, index));
+            nodeList[index].reset(new roscan::RosCanNode{name, index});
             nodeList[index]->start();
 
             // if successful, create thread to loop spinOnce for any subscribers
@@ -41,11 +41,11 @@ uint8_t RosCanNodeManager::registerNode(std::string& name, uint8_t hashName) {
     // TODO send response?
 }
 
-void RosCanNodeManager::deregisterNode(uint8_t id) {
+void RosCanNodeManager::deregisterNode(const uint8_t id) {
     {
-        boost::mutex::scoped_lock nodeListLock(nodeListMutex);
-        nodeList[id]->shutdown(); // may leak memory!!!
-        nodeList[id] = nullptr;
+        std::lock_guard<std::mutex> nodeListLock{nodeListMutex};
+        nodeList[id]->shutdown();
+        nodeList[id].reset();
     }
 
     //TODO send response, will need to store nodeid if so.
