@@ -32,7 +32,6 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "common.h"
 #include "subscription.h"
 #include "RosCanNode.hpp"
 #include "connection_manager.h"
@@ -45,11 +44,7 @@
 #include "intraprocess_subscriber_link.h"
 #include "callback_queue_interface.h"
 #include "subscription_queue.h"
-#include <cerrno>
-#include <fcntl.h>
 #include <ros/connection.h>
-#include <ros/file_log.h>
-#include <ros/io.h>
 #include <ros/message_deserializer.h>
 #include <ros/subscription_callback_helper.h>
 #include <ros/transport/transport_tcp.h>
@@ -60,9 +55,6 @@ using XmlRpc::XmlRpcValue;
 
 namespace roscan {
 
-Subscription::Subscription(const RosCanNodePtr& node, const std::string& name, const std::string& md5sum, const std::string& datatype, const ros::TransportHints& transport_hints)
-    : name_(name), md5sum_(md5sum), datatype_(datatype), nonconst_callbacks_(0), dropped_(false), shutting_down_(false), transport_hints_(transport_hints), node_(node) {}
-
 Subscription::~Subscription() {
     pending_connections_.clear();
     callbacks_.clear();
@@ -70,10 +62,9 @@ Subscription::~Subscription() {
 
 void Subscription::shutdown() {
     {
-        boost::mutex::scoped_lock lock(shutdown_mutex_);
+        std::lock_guard<std::mutex> lock{shutdown_mutex_};
         shutting_down_ = true;
     }
-
     drop();
 }
 
@@ -83,18 +74,17 @@ XmlRpcValue Subscription::getStats() {
     XmlRpcValue conn_data;
     conn_data.setSize(0);
 
-    boost::mutex::scoped_lock lock(publisher_links_mutex_);
+    std::lock_guard<std::mutex> lock{publisher_links_mutex_};
 
     uint32_t cidx = 0;
-    for (V_PublisherLink::iterator c = publisher_links_.begin(); c != publisher_links_.end(); ++c) {
-        const PublisherLink::Stats& s = (*c)->getStats();
-        conn_data[cidx][0] = (*c)->getConnectionID();
+    for (const auto& c: publisher_links_) {
+        const auto& s = c->getStats();
+        conn_data[cidx][0] = c->getConnectionID();
         conn_data[cidx][1] = (int)s.bytes_received_;
         conn_data[cidx][2] = (int)s.messages_received_;
         conn_data[cidx][3] = (int)s.drops_;
         conn_data[cidx][4] = 0; // figure out something for this. not sure.
     }
-
     stats[1] = conn_data;
     return stats;
 }
@@ -102,23 +92,23 @@ XmlRpcValue Subscription::getStats() {
 // [(connection_id, publisher_xmlrpc_uri, direction, transport, topic_name, connected, connection_info_string)*]
 // e.g. [(1, 'http://host:54893/', 'i', 'TCPROS', '/chatter', 1, 'TCPROS connection on port 59746 to [host:34318 on socket 11]')]
 void Subscription::getInfo(XmlRpc::XmlRpcValue& info) {
-    boost::mutex::scoped_lock lock(publisher_links_mutex_);
+    std::lock_guard<std::mutex> lock{publisher_links_mutex_};
 
-    for (V_PublisherLink::iterator c = publisher_links_.begin(); c != publisher_links_.end(); ++c) {
+    for (const auto& c: publisher_links_) {
         XmlRpcValue curr_info;
-        curr_info[0] = (int)(*c)->getConnectionID();
-        curr_info[1] = (*c)->getPublisherXMLRPCURI();
+        curr_info[0] = (int)c->getConnectionID();
+        curr_info[1] = c->getPublisherXMLRPCURI();
         curr_info[2] = "i";
-        curr_info[3] = (*c)->getTransportType();
+        curr_info[3] = c->getTransportType();
         curr_info[4] = name_;
         curr_info[5] = true; // For length compatibility with rospy
-        curr_info[6] = (*c)->getTransportInfo();
+        curr_info[6] = c->getTransportInfo();
         info[info.size()] = curr_info;
     }
 }
 
 uint32_t Subscription::getNumPublishers() {
-    boost::mutex::scoped_lock lock(publisher_links_mutex_);
+    std::lock_guard<std::mutex> lock{publisher_links_mutex_};
     return (uint32_t)publisher_links_.size();
 }
 
@@ -135,27 +125,23 @@ void Subscription::dropAllConnections() {
     V_PublisherLink localsubscribers;
 
     {
-        boost::mutex::scoped_lock lock(publisher_links_mutex_);
+        std::lock_guard<std::mutex> lock{publisher_links_mutex_};
         localsubscribers.swap(publisher_links_);
     }
 
-    V_PublisherLink::iterator it = localsubscribers.begin();
-    V_PublisherLink::iterator end = localsubscribers.end();
-    for (; it != end; ++it) {
-        (*it)->drop();
+    for (const auto& s: localsubscribers) {
+        s->drop();
     }
 }
 
 void Subscription::addLocalConnection(const PublicationPtr& pub) {
-    boost::mutex::scoped_lock lock(publisher_links_mutex_);
+    std::lock_guard<std::mutex> lock{publisher_links_mutex_};
     if (dropped_) {
         return;
     }
 
-    ROSCPP_LOG_DEBUG("Creating intraprocess link for topic [%s]", name_.c_str());
-
-    IntraProcessPublisherLinkPtr pub_link(boost::make_shared<IntraProcessPublisherLink>(node_, shared_from_this(), node_->xmlrpc_manager()->getServerURI(), transport_hints_));
-    IntraProcessSubscriberLinkPtr sub_link(boost::make_shared<IntraProcessSubscriberLink>(node_, pub));
+    IntraProcessPublisherLinkPtr pub_link{boost::make_shared<IntraProcessPublisherLink>(node_, shared_from_this(), node_->xmlrpc_manager()->getServerURI(), transport_hints_)};
+    IntraProcessSubscriberLinkPtr sub_link{boost::make_shared<IntraProcessSubscriberLink>(node_, pub)};
     pub_link->setPublisher(sub_link);
     sub_link->setSubscriber(pub_link);
 
@@ -171,109 +157,92 @@ bool urisEqual(const std::string& uri1, const std::string& uri2) {
     return port1 == port2 && host1 == host2;
 }
 
-bool Subscription::pubUpdate(const V_string& new_pubs) {
-    boost::mutex::scoped_lock lock(shutdown_mutex_);
+bool Subscription::pubUpdate(const std::vector<std::string>& new_pubs) {
+    std::lock_guard<std::mutex> lock{shutdown_mutex_};
 
     if (shutting_down_ || dropped_) {
         return false;
     }
 
-    bool retval = true;
+    auto retval = true;
 
     {
         std::stringstream ss;
 
-        for (V_string::const_iterator up_i = new_pubs.begin(); up_i != new_pubs.end(); ++up_i) {
-            ss << *up_i << ", ";
+        for (const auto& up: new_pubs) {
+            ss << up << ", ";
         }
 
         ss << " already have these connections: ";
         {
-            boost::mutex::scoped_lock lock(publisher_links_mutex_);
-            for (V_PublisherLink::iterator spc = publisher_links_.begin(); spc != publisher_links_.end(); ++spc) {
-                ss << (*spc)->getPublisherXMLRPCURI() << ", ";
+            std::lock_guard<std::mutex> lock{publisher_links_mutex_};
+            for (const auto& spc: publisher_links_) {
+                ss << spc->getPublisherXMLRPCURI() << ", ";
             }
         }
 
-        boost::mutex::scoped_lock lock(pending_connections_mutex_);
-        S_PendingConnection::iterator it = pending_connections_.begin();
-        S_PendingConnection::iterator end = pending_connections_.end();
-        for (; it != end; ++it) {
-            ss << (*it)->getRemoteURI() << ", ";
+        std::lock_guard<std::mutex> lock{pending_connections_mutex_};
+        for (const auto& c: pending_connections_) {
+            ss << c->getRemoteURI() << ", ";
         }
-
-        ROSCPP_LOG_DEBUG("Publisher update for [%s]: %s", name_.c_str(), ss.str().c_str());
     }
 
-    V_string additions;
+    std::vector<std::string> additions;
     V_PublisherLink subtractions;
     V_PublisherLink to_add;
     // could use the STL set operations... but these sets are so small
     // it doesn't really matter.
     {
-        boost::mutex::scoped_lock lock(publisher_links_mutex_);
+        std::lock_guard<std::mutex> lock{publisher_links_mutex_};
 
-        for (V_PublisherLink::iterator spc = publisher_links_.begin(); spc != publisher_links_.end(); ++spc) {
-            bool found = false;
-            for (V_string::const_iterator up_i = new_pubs.begin(); !found && up_i != new_pubs.end(); ++up_i) {
-                if (urisEqual((*spc)->getPublisherXMLRPCURI(), *up_i)) {
+        for (const auto& spc: publisher_links_) {
+            auto found = false;
+            for (const auto& up: new_pubs) {
+                if (urisEqual(spc->getPublisherXMLRPCURI(), up)) {
                     found = true;
                     break;
                 }
             }
-
             if (!found) {
-                subtractions.push_back(*spc);
+                subtractions.push_back(spc);
             }
         }
 
-        for (V_string::const_iterator up_i = new_pubs.begin(); up_i != new_pubs.end(); ++up_i) {
-            bool found = false;
-            for (V_PublisherLink::iterator spc = publisher_links_.begin(); !found && spc != publisher_links_.end(); ++spc) {
-                if (urisEqual(*up_i, (*spc)->getPublisherXMLRPCURI())) {
+        for (const auto& up: new_pubs) {
+            auto found = false;
+            for (const auto& spc: publisher_links_) {
+                if (urisEqual(up, spc->getPublisherXMLRPCURI())) {
                     found = true;
                     break;
                 }
             }
-
             if (!found) {
-                boost::mutex::scoped_lock lock(pending_connections_mutex_);
-                S_PendingConnection::iterator it = pending_connections_.begin();
-                S_PendingConnection::iterator end = pending_connections_.end();
-                for (; it != end; ++it) {
-                    if (urisEqual(*up_i, (*it)->getRemoteURI())) {
+                std::lock_guard<std::mutex> lock{pending_connections_mutex_};
+                for (const auto& c: pending_connections_) {
+                    if (urisEqual(up, c->getRemoteURI())) {
                         found = true;
                         break;
                     }
                 }
             }
-
             if (!found) {
-                additions.push_back(*up_i);
+                additions.push_back(up);
             }
         }
     }
 
-    for (V_PublisherLink::iterator i = subtractions.begin(); i != subtractions.end(); ++i) {
-        const PublisherLinkPtr& link = *i;
+    for (const auto& link: subtractions) {
         if (link->getPublisherXMLRPCURI() != node_->xmlrpc_manager()->getServerURI()) {
-            ROSCPP_LOG_DEBUG("Disconnecting from publisher [%s] of topic [%s] at [%s]",
-                             link->getCallerID().c_str(), name_.c_str(), link->getPublisherXMLRPCURI().c_str());
             link->drop();
-        } else {
-            ROSCPP_LOG_DEBUG("Disconnect: skipping myself for topic [%s]", name_.c_str());
         }
     }
 
-    for (V_string::iterator i = additions.begin(); i != additions.end(); ++i) {
+    for (const auto& s: additions) {
         // this function should never negotiate a self-subscription
-        if (node_->xmlrpc_manager()->getServerURI() != *i) {
-            retval &= negotiateConnection(*i);
-        } else {
-            ROSCPP_LOG_DEBUG("Skipping myself (%s, %s)", name_.c_str(), node_->xmlrpc_manager()->getServerURI().c_str());
+        if (node_->xmlrpc_manager()->getServerURI() != s) {
+            retval &= negotiateConnection(s);
         }
     }
-
     return retval;
 }
 
@@ -281,21 +250,22 @@ bool Subscription::negotiateConnection(const std::string& xmlrpc_uri) {
     XmlRpcValue tcpros_array, protos_array, params;
     XmlRpcValue udpros_array;
     ros::TransportUDPPtr udp_transport;
-    int protos = 0;
-    V_string transports = transport_hints_.getTransports();
+    auto protos = 0;
+    auto transports = transport_hints_.getTransports();
     if (transports.empty()) {
         transport_hints_.reliable();
         transports = transport_hints_.getTransports();
     }
-    for (V_string::const_iterator it = transports.begin(); it != transports.end(); ++it) {
-        if (*it == "UDP") {
-            int max_datagram_size = transport_hints_.getMaxDatagramSize();
+    for (const auto& t: transports) {
+        if (t == "UDP") {
+            auto max_datagram_size = transport_hints_.getMaxDatagramSize();
             udp_transport = boost::make_shared<ros::TransportUDP>(&node_->poll_manager()->getPollSet());
-            if (!max_datagram_size)
+            if (!max_datagram_size) {
                 max_datagram_size = udp_transport->getMaxDatagramSize();
+            }
             udp_transport->createIncoming(0, false);
             udpros_array[0] = "UDPROS";
-            M_string m;
+            std::map<std::string, std::string> m;
             m["topic"] = getName();
             m["md5sum"] = md5sum();
             m["callerid"] = node_->getName();
@@ -310,11 +280,11 @@ bool Subscription::negotiateConnection(const std::string& xmlrpc_uri) {
             udpros_array[4] = max_datagram_size;
 
             protos_array[protos++] = udpros_array;
-        } else if (*it == "TCP") {
-            tcpros_array[0] = std::string("TCPROS");
+        } else if (t == "TCP") {
+            tcpros_array[0] = std::string{"TCPROS"};
             protos_array[protos++] = tcpros_array;
         } else {
-            ROS_WARN("Unsupported transport type hinted: %s, skipping", it->c_str());
+            ROS_WARN("Unsupported transport type hinted: %s, skipping", t.c_str());
         }
     }
     params[0] = node_->getName();
@@ -327,34 +297,27 @@ bool Subscription::negotiateConnection(const std::string& xmlrpc_uri) {
         return false;
     }
 
-    XmlRpc::XmlRpcClient* c = new XmlRpc::XmlRpcClient(peer_host.c_str(),
-                                                       peer_port, "/");
+    XmlRpc::XmlRpcClient *c = new XmlRpc::XmlRpcClient{peer_host.c_str(), (int)peer_port, "/"};
 
     // Initiate the negotiation.  We'll come back and check on it later.
     if (!c->executeNonBlock("requestTopic", params)) {
-        ROSCPP_LOG_DEBUG("Failed to contact publisher [%s:%d] for topic [%s]",
-                         peer_host.c_str(), peer_port, name_.c_str());
         delete c;
         if (udp_transport) {
             udp_transport->close();
         }
-
         return false;
     }
 
-    ROSCPP_LOG_DEBUG("Began asynchronous xmlrpc connection to [%s:%d]", peer_host.c_str(), peer_port);
-
     // The PendingConnectionPtr takes ownership of c, and will delete it on
     // destruction.
-    PendingConnectionPtr conn(boost::make_shared<PendingConnection>(c, udp_transport, shared_from_this(), xmlrpc_uri));
+    PendingConnectionPtr conn{boost::make_shared<PendingConnection>(c, udp_transport, shared_from_this(), xmlrpc_uri)};
 
     node_->xmlrpc_manager()->addASyncConnection(conn);
     // Put this connection on the list that we'll look at later.
     {
-        boost::mutex::scoped_lock pending_connections_lock(pending_connections_mutex_);
+        std::lock_guard<std::mutex> pending_connections_lock{pending_connections_mutex_};
         pending_connections_.insert(conn);
     }
-
     return true;
 }
 
@@ -365,78 +328,65 @@ void closeTransport(const ros::TransportUDPPtr& trans) {
 }
 
 void Subscription::pendingConnectionDone(const PendingConnectionPtr& conn, XmlRpcValue& result) {
-    boost::mutex::scoped_lock lock(shutdown_mutex_);
+    std::lock_guard<std::mutex> lock{shutdown_mutex_};
     if (shutting_down_ || dropped_) {
         return;
     }
 
     {
-        boost::mutex::scoped_lock pending_connections_lock(pending_connections_mutex_);
+        std::lock_guard<std::mutex> pending_connections_lock{pending_connections_mutex_};
         pending_connections_.erase(conn);
     }
 
     ros::TransportUDPPtr udp_transport;
 
-    std::string peer_host = conn->getClient()->getHost();
-    uint32_t peer_port = conn->getClient()->getPort();
+    const auto peer_host = conn->getClient()->getHost();
+    const auto peer_port = conn->getClient()->getPort();
     std::stringstream ss;
     ss << "http://" << peer_host << ":" << peer_port << "/";
-    std::string xmlrpc_uri = ss.str();
+    const auto xmlrpc_uri = ss.str();
     udp_transport = conn->getUDPTransport();
 
     XmlRpc::XmlRpcValue proto;
     if (!node_->xmlrpc_manager()->validateXmlrpcResponse("requestTopic", result, proto)) {
-        ROSCPP_LOG_DEBUG("Failed to contact publisher [%s:%d] for topic [%s]", peer_host.c_str(), peer_port, name_.c_str());
         closeTransport(udp_transport);
         return;
     }
 
     if (proto.size() == 0) {
-        ROSCPP_LOG_DEBUG("Couldn't agree on any common protocols with [%s] for topic [%s]", xmlrpc_uri.c_str(), name_.c_str());
         closeTransport(udp_transport);
         return;
     }
 
     if (proto.getType() != XmlRpcValue::TypeArray) {
-        ROSCPP_LOG_DEBUG("Available protocol info returned from %s is not a list.", xmlrpc_uri.c_str());
         closeTransport(udp_transport);
         return;
     }
     if (proto[0].getType() != XmlRpcValue::TypeString) {
-        ROSCPP_LOG_DEBUG("Available protocol info list doesn't have a string as its first element.");
         closeTransport(udp_transport);
         return;
     }
 
     std::string proto_name = proto[0];
     if (proto_name == "TCPROS") {
-        if (proto.size() != 3 ||
-            proto[1].getType() != XmlRpcValue::TypeString ||
-            proto[2].getType() != XmlRpcValue::TypeInt) {
-            ROSCPP_LOG_DEBUG("publisher implements TCPROS, but the "
-                             "parameters aren't string,int");
+        if (proto.size() != 3 || proto[1].getType() != XmlRpcValue::TypeString || proto[2].getType() != XmlRpcValue::TypeInt) {
             return;
         }
         std::string pub_host = proto[1];
         int pub_port = proto[2];
-        ROSCPP_CONN_LOG_DEBUG("Connecting via tcpros to topic [%s] at host [%s:%d]", name_.c_str(), pub_host.c_str(), pub_port);
 
-        ros::TransportTCPPtr transport(boost::make_shared<ros::TransportTCP>(&node_->poll_manager()->getPollSet()));
+        ros::TransportTCPPtr transport{boost::make_shared<ros::TransportTCP>(&node_->poll_manager()->getPollSet())};
         if (transport->connect(pub_host, pub_port)) {
-            ros::ConnectionPtr connection(boost::make_shared<ros::Connection>());
-            TransportPublisherLinkPtr pub_link(boost::make_shared<TransportPublisherLink>(node_, shared_from_this(), xmlrpc_uri, transport_hints_));
+            ros::ConnectionPtr connection{boost::make_shared<ros::Connection>()};
+            TransportPublisherLinkPtr pub_link{boost::make_shared<TransportPublisherLink>(node_, shared_from_this(), xmlrpc_uri, transport_hints_)};
 
             connection->initialize(transport, false, ros::HeaderReceivedFunc());
             pub_link->initialize(connection);
 
             node_->connection_manager()->addConnection(connection);
 
-            boost::mutex::scoped_lock lock(publisher_links_mutex_);
+            std::lock_guard<std::mutex> lock{publisher_links_mutex_};
             addPublisherLink(pub_link);
-
-            ROSCPP_CONN_LOG_DEBUG("Connected to publisher of topic [%s] at [%s:%d]", name_.c_str(), pub_host.c_str(), pub_port);
-        } else {
-            ROSCPP_CONN_LOG_DEBUG("Failed to connect to publisher of topic [%s] at [%s:%d]", name_.c_str(), pub_host.c_str(), pub_port);
         }
     } else if (proto_name == "UDPROS") {
         if (proto.size() != 6 ||
@@ -445,8 +395,6 @@ void Subscription::pendingConnectionDone(const PendingConnectionPtr& conn, XmlRp
             proto[3].getType() != XmlRpcValue::TypeInt ||
             proto[4].getType() != XmlRpcValue::TypeInt ||
             proto[5].getType() != XmlRpcValue::TypeBase64) {
-            ROSCPP_LOG_DEBUG("publisher implements UDPROS, but the "
-                             "parameters aren't string,int,int,int,base64");
             closeTransport(udp_transport);
             return;
         }
@@ -455,49 +403,41 @@ void Subscription::pendingConnectionDone(const PendingConnectionPtr& conn, XmlRp
         int conn_id = proto[3];
         int max_datagram_size = proto[4];
         std::vector<char> header_bytes = proto[5];
-        boost::shared_array<uint8_t> buffer(new uint8_t[header_bytes.size()]);
+        boost::shared_array<uint8_t> buffer{new uint8_t[header_bytes.size()]};
         memcpy(buffer.get(), &header_bytes[0], header_bytes.size());
         ros::Header h;
         std::string err;
         if (!h.parse(buffer, header_bytes.size(), err)) {
-            ROSCPP_LOG_DEBUG("Unable to parse UDPROS connection header: %s", err.c_str());
             closeTransport(udp_transport);
             return;
         }
-        ROSCPP_LOG_DEBUG("Connecting via udpros to topic [%s] at host [%s:%d] connection id [%08x] max_datagram_size [%d]", name_.c_str(), pub_host.c_str(), pub_port, conn_id, max_datagram_size);
 
         std::string error_msg;
         if (h.getValue("error", error_msg)) {
-            ROSCPP_LOG_DEBUG("Received error message in header for connection to [%s]: [%s]", xmlrpc_uri.c_str(), error_msg.c_str());
             closeTransport(udp_transport);
             return;
         }
 
-        TransportPublisherLinkPtr pub_link(boost::make_shared<TransportPublisherLink>(node_, shared_from_this(), xmlrpc_uri, transport_hints_));
+        TransportPublisherLinkPtr pub_link{boost::make_shared<TransportPublisherLink>(node_, shared_from_this(), xmlrpc_uri, transport_hints_)};
         if (pub_link->setHeader(h)) {
-            ros::ConnectionPtr connection(boost::make_shared<ros::Connection>());
+            ros::ConnectionPtr connection{boost::make_shared<ros::Connection>()};
             connection->initialize(udp_transport, false, NULL);
             connection->setHeader(h);
             pub_link->initialize(connection);
 
             node_->connection_manager()->addConnection(connection);
 
-            boost::mutex::scoped_lock lock(publisher_links_mutex_);
+            std::lock_guard<std::mutex> lock{publisher_links_mutex_};
             addPublisherLink(pub_link);
-
-            ROSCPP_LOG_DEBUG("Connected to publisher of topic [%s] at [%s:%d]", name_.c_str(), pub_host.c_str(), pub_port);
         } else {
-            ROSCPP_LOG_DEBUG("Failed to connect to publisher of topic [%s] at [%s:%d]", name_.c_str(), pub_host.c_str(), pub_port);
             closeTransport(udp_transport);
             return;
         }
-    } else {
-        ROSCPP_LOG_DEBUG("Publisher offered unsupported transport [%s]", proto_name.c_str());
     }
 }
 
-uint32_t Subscription::handleMessage(const ros::SerializedMessage& m, bool ser, bool nocopy, const boost::shared_ptr<M_string>& connection_header, const PublisherLinkPtr& link) {
-    boost::mutex::scoped_lock lock(callbacks_mutex_);
+uint32_t Subscription::handleMessage(const ros::SerializedMessage& m, const bool ser, const bool nocopy, const boost::shared_ptr<std::map<std::string, std::string>>& connection_header, const PublisherLinkPtr& link) {
+    std::lock_guard<std::mutex> lock{callbacks_mutex_};
 
     uint32_t drops = 0;
 
@@ -506,23 +446,17 @@ uint32_t Subscription::handleMessage(const ros::SerializedMessage& m, bool ser, 
     // garbage to the messages with different C++ types than the first one.
     cached_deserializers_.clear();
 
-    ros::Time receipt_time = ros::Time::now();
+    const auto receipt_time = ros::Time::now();
 
-    for (V_CallbackInfo::iterator cb = callbacks_.begin(); cb != callbacks_.end(); ++cb) {
-        const CallbackInfoPtr& info = *cb;
-
-        ROS_ASSERT(info->callback_queue_);
-
-        const std::type_info* ti = &info->helper_->getTypeInfo();
+    for (const auto& info: callbacks_) {
+        const std::type_info *ti = &info->helper_->getTypeInfo();
 
         if ((nocopy && m.type_info && *ti == *m.type_info) || (ser && (!m.type_info || *ti != *m.type_info))) {
             ros::MessageDeserializerPtr deserializer;
 
-            V_TypeAndDeserializer::iterator des_it = cached_deserializers_.begin();
-            V_TypeAndDeserializer::iterator des_end = cached_deserializers_.end();
-            for (; des_it != des_end; ++des_it) {
-                if (*des_it->first == *ti) {
-                    deserializer = des_it->second;
+            for (const auto& des: cached_deserializers_) {
+                if (*des.first == *ti) {
+                    deserializer = des.second;
                     break;
                 }
             }
@@ -532,8 +466,8 @@ uint32_t Subscription::handleMessage(const ros::SerializedMessage& m, bool ser, 
                 cached_deserializers_.push_back(std::make_pair(ti, deserializer));
             }
 
-            bool was_full = false;
-            bool nonconst_need_copy = false;
+            auto was_full = false;
+            auto nonconst_need_copy = false;
             if (callbacks_.size() > 1) {
                 nonconst_need_copy = true;
             }
@@ -560,21 +494,16 @@ uint32_t Subscription::handleMessage(const ros::SerializedMessage& m, bool ser, 
         li.receipt_time = receipt_time;
         latched_messages_[link] = li;
     }
-
     cached_deserializers_.clear();
-
     return drops;
 }
 
-bool Subscription::addCallback(const ros::SubscriptionCallbackHelperPtr& helper, const std::string& md5sum, CallbackQueueInterface* queue, int32_t queue_size, const ros::VoidConstPtr& tracked_object, bool allow_concurrent_callbacks) {
-    ROS_ASSERT(helper);
-    ROS_ASSERT(queue);
-
+bool Subscription::addCallback(const ros::SubscriptionCallbackHelperPtr& helper, const std::string& md5sum, CallbackQueueInterface *const queue, const int32_t queue_size, const ros::VoidConstPtr& tracked_object, const bool allow_concurrent_callbacks) {
     //statistics_.init(helper);
 
     // Decay to a real type as soon as we have a subscriber with a real type
     {
-        boost::mutex::scoped_lock lock(md5sum_mutex_);
+        std::lock_guard<std::mutex> lock{md5sum_mutex_};
         if (md5sum_ == "*" && md5sum != "*") {
             md5sum_ = md5sum;
         }
@@ -585,9 +514,9 @@ bool Subscription::addCallback(const ros::SubscriptionCallbackHelperPtr& helper,
     }
 
     {
-        boost::mutex::scoped_lock lock(callbacks_mutex_);
+        std::lock_guard<std::mutex> lock{callbacks_mutex_};
 
-        CallbackInfoPtr info(boost::make_shared<CallbackInfo>());
+        CallbackInfoPtr info{boost::make_shared<CallbackInfo>()};
         info->helper_ = helper;
         info->callback_queue_ = queue;
         info->subscription_queue_ = boost::make_shared<SubscriptionQueue>(name_, queue_size, allow_concurrent_callbacks);
@@ -606,19 +535,15 @@ bool Subscription::addCallback(const ros::SubscriptionCallbackHelperPtr& helper,
 
         // if we have any latched links, we need to immediately schedule callbacks
         if (!latched_messages_.empty()) {
-            boost::mutex::scoped_lock lock(publisher_links_mutex_);
+            std::lock_guard<std::mutex> lock{publisher_links_mutex_};
 
-            V_PublisherLink::iterator it = publisher_links_.begin();
-            V_PublisherLink::iterator end = publisher_links_.end();
-            for (; it != end; ++it) {
-                const PublisherLinkPtr& link = *it;
+            for (const auto& link: publisher_links_) {
                 if (link->isLatched()) {
-                    M_PublisherLinkToLatchInfo::iterator des_it = latched_messages_.find(link);
+                    const auto des_it = latched_messages_.find(link);
                     if (des_it != latched_messages_.end()) {
                         const LatchInfo& latch_info = des_it->second;
-
-                        ros::MessageDeserializerPtr des(boost::make_shared<ros::MessageDeserializer>(helper, latch_info.message, latch_info.connection_header));
-                        bool was_full = false;
+                        ros::MessageDeserializerPtr des{boost::make_shared<ros::MessageDeserializer>(helper, latch_info.message, latch_info.connection_header)};
+                        auto was_full = false;
                         info->subscription_queue_->push(info->helper_, des, info->has_tracked_object_, info->tracked_object_, true, latch_info.receipt_time, &was_full);
                         if (!was_full) {
                             info->callback_queue_->addCallback(info->subscription_queue_, (uint64_t)info.get());
@@ -628,25 +553,20 @@ bool Subscription::addCallback(const ros::SubscriptionCallbackHelperPtr& helper,
             }
         }
     }
-
     return true;
 }
 
 void Subscription::removeCallback(const ros::SubscriptionCallbackHelperPtr& helper) {
     CallbackInfoPtr info;
     {
-        boost::mutex::scoped_lock cbs_lock(callbacks_mutex_);
-        for (V_CallbackInfo::iterator it = callbacks_.begin(); it != callbacks_.end(); ++it) {
-            if ((*it)->helper_ == helper) {
+        std::lock_guard<std::mutex> cbs_lock{callbacks_mutex_};
+        const auto it = std::find_if(callbacks_.cbegin(), callbacks_.cend(), [&helper](const auto& c){ return helper == c->helper_; });
+        if (it != callbacks_.cend()) {
                 info = *it;
                 callbacks_.erase(it);
-
                 if (!helper->isConst()) {
                     --nonconst_callbacks_;
                 }
-
-                break;
-            }
         }
     }
 
@@ -658,21 +578,17 @@ void Subscription::removeCallback(const ros::SubscriptionCallbackHelperPtr& help
 
 void Subscription::headerReceived(const PublisherLinkPtr& link, const ros::Header& h) {
     (void)h;
-    boost::mutex::scoped_lock lock(md5sum_mutex_);
+    std::lock_guard<std::mutex> lock{md5sum_mutex_};
     if (md5sum_ == "*") {
         md5sum_ = link->getMD5Sum();
     }
 }
 
-void Subscription::addPublisherLink(const PublisherLinkPtr& link) {
-    publisher_links_.push_back(link);
-}
-
 void Subscription::removePublisherLink(const PublisherLinkPtr& pub_link) {
-    boost::mutex::scoped_lock lock(publisher_links_mutex_);
+    std::lock_guard<std::mutex> lock{publisher_links_mutex_};
 
-    V_PublisherLink::iterator it = std::find(publisher_links_.begin(), publisher_links_.end(), pub_link);
-    if (it != publisher_links_.end()) {
+    const auto it = std::find(publisher_links_.cbegin(), publisher_links_.cend(), pub_link);
+    if (it != publisher_links_.cend()) {
         publisher_links_.erase(it);
     }
 
@@ -682,9 +598,8 @@ void Subscription::removePublisherLink(const PublisherLinkPtr& pub_link) {
 }
 
 void Subscription::getPublishTypes(bool& ser, bool& nocopy, const std::type_info& ti) {
-    boost::mutex::scoped_lock lock(callbacks_mutex_);
-    for (V_CallbackInfo::iterator cb = callbacks_.begin(); cb != callbacks_.end(); ++cb) {
-        const CallbackInfoPtr& info = *cb;
+    std::lock_guard<std::mutex> lock{callbacks_mutex_};
+    for (const auto& info: callbacks_) {
         if (info->helper_->getTypeInfo() == ti) {
             nocopy = true;
         } else {
@@ -697,12 +612,8 @@ void Subscription::getPublishTypes(bool& ser, bool& nocopy, const std::type_info
     }
 }
 
-const std::string Subscription::datatype() {
-    return datatype_;
-}
-
 const std::string Subscription::md5sum() {
-    boost::mutex::scoped_lock lock(md5sum_mutex_);
+    std::lock_guard<std::mutex> lock{md5sum_mutex_};
     return md5sum_;
 }
 
