@@ -29,7 +29,11 @@
 #define ROSCAN_TRANSPORT_SUBSCRIBER_LINK_H
 
 #include "common.h"
+#include "rosdefs.h"
 #include "subscriber_link.h"
+#include <ros/transport/transport.h>
+#include <boost/signals2/connection.hpp>
+#include <queue>
 
 namespace roscan {
 
@@ -37,25 +41,25 @@ namespace roscan {
 class TransportSubscriberLink : public SubscriberLink {
     public:
         TransportSubscriberLink(const RosCanNodePtr& node)
-            : SubscriberLink(node), writing_message_(false), header_written_(false), queue_full_(false) {}
-        virtual ~TransportSubscriberLink() { drop(); }
+            : SubscriberLink{node}, writing_message_{false}, header_written_{false}, queue_full_{false} {}
+        ~TransportSubscriberLink() override { drop(); }
 
         bool initialize(const ros::ConnectionPtr& connection);
         bool handleHeader(const ros::Header& header);
 
-        const ros::ConnectionPtr& getConnection() { return connection_; }
+        const ros::ConnectionPtr& getConnection() const { return connection_; }
 
-        virtual void enqueueMessage(const ros::SerializedMessage& m, bool ser, bool nocopy);
-        virtual void drop();
-        virtual std::string getTransportType();
-        virtual std::string getTransportInfo();
+        void enqueueMessage(const ros::SerializedMessage& m, const bool ser, const bool nocopy) override;
+        void drop() override;
+        std::string getTransportType() override { return connection_->getTransport()->getType(); }
+        std::string getTransportInfo() override { return connection_->getTransport()->getTransportInfo(); }
 
     private:
         void onConnectionDropped(const ros::ConnectionPtr& conn);
 
         void onHeaderWritten(const ros::ConnectionPtr& conn);
         void onMessageWritten(const ros::ConnectionPtr& conn);
-        void startMessageWrite(bool immediate_write);
+        void startMessageWrite(const bool immediate_write);
 
         bool writing_message_;
         bool header_written_;
@@ -64,7 +68,7 @@ class TransportSubscriberLink : public SubscriberLink {
         boost::signals2::connection dropped_conn_;
 
         std::queue<ros::SerializedMessage> outbox_;
-        boost::mutex outbox_mutex_;
+        std::mutex outbox_mutex_;
         bool queue_full_;
 };
 
