@@ -11,24 +11,24 @@ MessageBuffer& MessageBuffer::instance() {
 }
 
 void MessageBuffer::push(const can_frame& data) {
-    std::unique_lock<std::mutex> lk(mutex);
+    std::unique_lock<std::mutex> lock{mutex};
 
     q.push(data);
     
-    lk.unlock();
+    lock.unlock();
     cv.notify_one();
 }
 
 MessageBuffer::MessageBuffer() {
     std::thread sender{[this]() {
         while (true) {
-            std::unique_lock<std::mutex> lk(mutex);
-            cv.wait(lk, [this](){ return q.size() > 0; });
+            std::unique_lock<std::mutex> lock{mutex};
+            cv.wait(lock, [this](){ return q.size() > 0; });
 
-            auto frame = q.front();
+            const auto frame = q.front();
             q.pop();
             
-            lk.unlock();
+            lock.unlock();
             
             // debug exit condition
             if (frame.__res0 == 8) {
@@ -36,16 +36,16 @@ MessageBuffer::MessageBuffer() {
                 break;
             }
 
-            std::cout << "header = " << frame.can_id << ", data = ";
+            printf("Sending header = %#08X, length = %d, data = ", frame.can_id, frame.can_dlc);
             for (int i = 0; i < frame.can_dlc;++i) {
-                printf("%02X", frame.data[i]);
+                printf("%02X ", frame.data[i]);
             }
-            std::cout << "\n";
+            printf("\n");
 
             // CAN port is assumed to be open
-            if (CANHelpers::send_can_port(&frame) < 0) {
-                std::cerr << "send failed: frame could not be sent\n";
-            }
+            // if (CANHelpers::send_can_port(frame) < 0) {
+            //     std::cerr << "send failed: frame could not be sent\n";
+            // }
         }
     }};
     sender.detach();
