@@ -32,7 +32,6 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "common.h"
 #include "RosCanNode.hpp"
 #include "publisher_link.h"
 #include "transport_publisher_link.h"
@@ -43,7 +42,6 @@
 #include "internal_timer_manager.h"
 #include "timer_manager.h"
 #include <ros/connection.h>
-#include <ros/file_log.h>
 #include <ros/header.h>
 #include <ros/transport/transport.h>
 #include <ros/transport/transport_tcp.h>
@@ -70,9 +68,9 @@ bool TransportPublisherLink::initialize(const ros::ConnectionPtr& connection) {
     if (connection_->getTransport()->requiresHeader()) {
         connection_->setHeaderReceivedCallback(boost::bind(&TransportPublisherLink::onHeaderReceived, this, _1, _2));
 
-        SubscriptionPtr parent = parent_.lock();
+        const auto parent = parent_.lock();
 
-        ros::M_string header;
+        std::map<std::string, std::string> header;
         header["topic"] = parent->getName();
         header["md5sum"] = parent->md5sum();
         header["callerid"] = node_->getName();
@@ -89,19 +87,13 @@ void TransportPublisherLink::drop() {
     dropping_ = true;
     connection_->drop(ros::Connection::Destructing);
 
-    if (SubscriptionPtr parent = parent_.lock()) {
+    if (const auto parent = parent_.lock()) {
         parent->removePublisherLink(shared_from_this());
     }
 }
 
-void TransportPublisherLink::onHeaderWritten(const ros::ConnectionPtr& conn) {
-    (void)conn;
-    // Do nothing
-}
-
 bool TransportPublisherLink::onHeaderReceived(const ros::ConnectionPtr& conn, const ros::Header& header) {
     (void)conn;
-    ROS_ASSERT(conn == connection_);
 
     if (!setHeader(header)) {
         drop();
@@ -117,7 +109,7 @@ bool TransportPublisherLink::onHeaderReceived(const ros::ConnectionPtr& conn, co
     return true;
 }
 
-void TransportPublisherLink::onMessageLength(const ros::ConnectionPtr& conn, const boost::shared_array<uint8_t>& buffer, uint32_t size, bool success) {
+void TransportPublisherLink::onMessageLength(const ros::ConnectionPtr& conn, const boost::shared_array<uint8_t>& buffer, const uint32_t size, const bool success) {
     (void)conn;
     (void)size;
     if (retry_timer_handle_ != -1) {
@@ -126,13 +118,11 @@ void TransportPublisherLink::onMessageLength(const ros::ConnectionPtr& conn, con
     }
 
     if (!success) {
-        if (connection_)
+        if (connection_) {
             connection_->read(4, boost::bind(&TransportPublisherLink::onMessageLength, this, _1, _2, _3, _4));
+        }
         return;
     }
-
-    ROS_ASSERT(conn == connection_);
-    ROS_ASSERT(size == 4);
 
     uint32_t len = *((uint32_t*)buffer.get());
 
@@ -147,14 +137,13 @@ void TransportPublisherLink::onMessageLength(const ros::ConnectionPtr& conn, con
     connection_->read(len, boost::bind(&TransportPublisherLink::onMessage, this, _1, _2, _3, _4));
 }
 
-void TransportPublisherLink::onMessage(const ros::ConnectionPtr& conn, const boost::shared_array<uint8_t>& buffer, uint32_t size, bool success) {
-    if (!success && !conn)
+void TransportPublisherLink::onMessage(const ros::ConnectionPtr& conn, const boost::shared_array<uint8_t>& buffer, const uint32_t size, const bool success) {
+    if (!success && !conn) {
         return;
-
-    ROS_ASSERT(conn == connection_);
+    }
 
     if (success) {
-        handleMessage(ros::SerializedMessage(buffer, size), true, false);
+        handleMessage(ros::SerializedMessage{buffer, size}, true, false);
     }
 
     if (success || !connection_->getTransport()->requiresHeader()) {
@@ -168,32 +157,27 @@ void TransportPublisherLink::onRetryTimer(const ros::SteadyTimerEvent&) {
     }
 
     if (needs_retry_ && ros::SteadyTime::now() > next_retry_) {
-        retry_period_ = std::min(retry_period_ * 2, ros::WallDuration(20));
+        retry_period_ = std::min(retry_period_ * 2, ros::WallDuration{20});
         needs_retry_ = false;
-        SubscriptionPtr parent = parent_.lock();
+        const auto parent = parent_.lock();
         // TODO: support retry on more than just TCP
         // For now, since UDP does not have a heartbeat, we do not attempt to retry
         // UDP connections since an error there likely means some invalid operation has
         // happened.
-        if (connection_->getTransport()->getType() == std::string("TCPROS")) {
-            std::string topic = parent ? parent->getName() : "unknown";
+        if (connection_->getTransport()->getType() == std::string{"TCPROS"}) {
+            const auto topic = parent ? parent->getName() : "unknown";
 
-            ros::TransportTCPPtr old_transport = boost::dynamic_pointer_cast<ros::TransportTCP>(connection_->getTransport());
-            ROS_ASSERT(old_transport);
-            const std::string& host = old_transport->getConnectedHost();
-            int port = old_transport->getConnectedPort();
+            const auto old_transport = boost::dynamic_pointer_cast<ros::TransportTCP>(connection_->getTransport());
+            const auto& host = old_transport->getConnectedHost();
+            const auto port = old_transport->getConnectedPort();
 
-            ROSCPP_CONN_LOG_DEBUG("Retrying connection to [%s:%d] for topic [%s]", host.c_str(), port, topic.c_str());
-
-            ros::TransportTCPPtr transport(boost::make_shared<ros::TransportTCP>(&node_->poll_manager()->getPollSet()));
+            ros::TransportTCPPtr transport{boost::make_shared<ros::TransportTCP>(&node_->poll_manager()->getPollSet())};
             if (transport->connect(host, port)) {
-                ros::ConnectionPtr connection(boost::make_shared<ros::Connection>());
+                ros::ConnectionPtr connection{boost::make_shared<ros::Connection>()};
                 connection->initialize(transport, false, ros::HeaderReceivedFunc());
                 initialize(connection);
 
                 node_->connection_manager()->addConnection(connection);
-            } else {
-                ROSCPP_CONN_LOG_DEBUG("connect() failed when retrying connection to [%s:%d] for topic [%s]", host.c_str(), port, topic.c_str());
             }
         } else if (parent) {
             parent->removePublisherLink(shared_from_this());
@@ -207,26 +191,21 @@ void TransportPublisherLink::onConnectionDropped(const ros::ConnectionPtr& conn,
         return;
     }
 
-    ROS_ASSERT(conn == connection_);
-
-    SubscriptionPtr parent = parent_.lock();
+    const auto parent = parent_.lock();
 
     if (reason == ros::Connection::TransportDisconnect) {
-        std::string topic = parent ? parent->getName() : "unknown";
+        const auto topic = parent ? parent->getName() : "unknown";
 
-        ROSCPP_CONN_LOG_DEBUG("Connection to publisher [%s] to topic [%s] dropped", connection_->getTransport()->getTransportInfo().c_str(), topic.c_str());
-
-        ROS_ASSERT(!needs_retry_);
         needs_retry_ = true;
         next_retry_ = ros::SteadyTime::now() + retry_period_;
 
         if (retry_timer_handle_ == -1) {
-            retry_period_ = ros::WallDuration(0.1);
+            retry_period_ = ros::WallDuration{0.1};
             next_retry_ = ros::SteadyTime::now() + retry_period_;
             // shared_from_this() shared_ptr is used to ensure TransportPublisherLink is not
             // destroyed in the middle of onRetryTimer execution
             // TODO RESTORE after callback_queue issue resolved
-            retry_timer_handle_ = getInternalTimerManager()->add(ros::WallDuration(retry_period_),
+            retry_timer_handle_ = getInternalTimerManager()->add(ros::WallDuration{retry_period_},
                                                                  boost::bind(&TransportPublisherLink::onRetryTimer, this, _1), node_->getInternalCallbackQueue().get(),
                                                                  shared_from_this(), false);
         } else {
@@ -237,23 +216,13 @@ void TransportPublisherLink::onConnectionDropped(const ros::ConnectionPtr& conn,
     }
 }
 
-void TransportPublisherLink::handleMessage(const ros::SerializedMessage& m, bool ser, bool nocopy) {
+void TransportPublisherLink::handleMessage(const ros::SerializedMessage& m, const bool ser, const bool nocopy) {
     stats_.bytes_received_ += m.num_bytes;
-    stats_.messages_received_++;
+    ++stats_.messages_received_;
 
-    SubscriptionPtr parent = parent_.lock();
-
-    if (parent) {
+    if (const auto parent = parent_.lock()) {
         stats_.drops_ += parent->handleMessage(m, ser, nocopy, getConnection()->getHeader().getValues(), shared_from_this());
     }
-}
-
-std::string TransportPublisherLink::getTransportType() {
-    return connection_->getTransport()->getType();
-}
-
-std::string TransportPublisherLink::getTransportInfo() {
-    return connection_->getTransport()->getTransportInfo();
 }
 
 } // namespace roscan
